Update to pdk-template Gemfile and puppet-module-gems
The PDK team is maintaining meta-gems (only containing dependencies)
for a number of core gems used by the PDK and the community at large.
See https://github.com/puppetlabs/puppet-module-gems for details

This commit switches to the pdk-templates core group/gem loop, updates
the location_for method, and adjusts the default dependencies to
use the puppet-module-gems.